@@ -18,7 +18,7 @@
         *************************************************************************************
     -->
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <lombok.version>1.18.20</lombok.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
