@@ -1,35 +1,36 @@
 package org.infinity.luix.utilities.serviceloader;
 
-import junit.framework.TestCase;
 import org.infinity.luix.utilities.serviceloader.testservice.SpiPrototypeInterface;
 import org.infinity.luix.utilities.serviceloader.testservice.SpiSingletonInterface;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-public class ServiceLoaderTests extends TestCase {
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ServiceLoaderTests  {
 
     @Test
     public void testSingletonInitialization() {
         // 单例模式下只会构造一次实例
-        Assert.assertEquals(1, ServiceLoader.forClass(SpiSingletonInterface.class)
-                .load("singleton").spiHello());
-        Assert.assertEquals(1, ServiceLoader.forClass(SpiSingletonInterface.class)
-                .load("singleton").spiHello());
+        assertThat(ServiceLoader.forClass(SpiSingletonInterface.class)
+                .load("singleton").spiHello()).isEqualTo(1);
+
+        assertThat(ServiceLoader.forClass(SpiSingletonInterface.class)
+                .load("singleton").spiHello()).isEqualTo(1);
     }
 
     @Test
     public void testPrototypeInitialization() {
         // 多例模式下在每次获取的时候进行实例化
-        Assert.assertEquals(1, ServiceLoader.forClass(SpiPrototypeInterface.class)
-                .load("prototype").spiHello());
-        Assert.assertEquals(2, ServiceLoader.forClass(SpiPrototypeInterface.class)
-                .load("prototype").spiHello());
+        assertThat(ServiceLoader.forClass(SpiPrototypeInterface.class)
+                .load("singleton").spiHello()).isEqualTo(1);
+        assertThat(ServiceLoader.forClass(SpiPrototypeInterface.class)
+                .load("singleton").spiHello()).isEqualTo(2);
     }
 
     @Test
     public void testChineseCharacterLoad() {
         // 单例模式下只会构造一次实例
-        Assert.assertEquals(1, ServiceLoader.forClass(SpiSingletonInterface.class)
-                .load("单例").spiHello());
+        assertThat(ServiceLoader.forClass(SpiSingletonInterface.class)
+                .load("单例").spiHello()).isEqualTo(1);
     }
 }
