@@ -1,206 +1,210 @@
 package org.infinity.luix.registry.consul;
 
+import com.ecwid.consul.v1.Response;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.collections4.MapUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.infinity.luix.core.registry.CommandFailbackAbstractRegistry;
 import org.infinity.luix.core.registry.listener.CommandListener;
 import org.infinity.luix.core.registry.listener.ProviderListener;
 import org.infinity.luix.core.server.listener.ConsumerProcessable;
 import org.infinity.luix.core.url.Url;
-import org.infinity.luix.registry.consul.client.AbstractConsulClient;
 import org.infinity.luix.registry.consul.utils.ConsulUtils;
-import org.infinity.luix.utilities.destory.Cleanable;
+import org.infinity.luix.utilities.destory.Destroyable;
 import org.infinity.luix.utilities.destory.ShutdownHook;
 
 import javax.annotation.concurrent.ThreadSafe;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
+import java.util.Set;
+import java.util.concurrent.*;
+
+import static org.infinity.luix.core.constant.RegistryConstants.DISCOVERY_INTERVAL;
+import static org.infinity.luix.core.constant.RegistryConstants.DISCOVERY_INTERVAL_VAL_DEFAULT;
 
 @Slf4j
 @ThreadSafe
-public class ConsulRegistry extends CommandFailbackAbstractRegistry implements Cleanable {
+public class ConsulRegistry extends CommandFailbackAbstractRegistry implements Destroyable {
 
     /**
-     * consul服务查询默认间隔时间。单位毫秒
+     * Consul client
      */
-    public static int DEFAULT_LOOKUP_INTERVAL = 30000;
+    private final LuixConsulClient                                                    consulClient;
+    /**
+     * Consul service instance status updater
+     */
+    private final ConsulStatusUpdater                                                 consulStatusUpdater;
+    /**
+     * Service discovery interval in milliseconds
+     */
+    private final int                                                                 discoverInterval;
+    /**
+     * Provider service notification thread pool
+     */
+    private final ThreadPoolExecutor                                                  notificationThreadPool;
+    /**
+     * Cache used to store provider urls
+     * Key:  form
+     * Value: protocol plus path to urls map
+     */
+    private final ConcurrentHashMap<String, ConcurrentHashMap<String, List<Url>>>     urlCache         = new ConcurrentHashMap<>();
+    /**
+     * Cache used to store commands
+     * Key: form
+     * Value: command string
+     */
+    private final ConcurrentHashMap<String, String>                                   commandCache     = new ConcurrentHashMap<>();
+    /**
+     * Key: form
+     * Value: lastConsulIndexId
+     */
+    private final ConcurrentHashMap<String, Long>                                     form2ConsulIndex = new ConcurrentHashMap<>();
+    /**
+     * Key: form
+     * Value: command string
+     */
+    private final ConcurrentHashMap<String, String>                                   form2Command     = new ConcurrentHashMap<>();
+    /**
+     * Key: protocol plus path
+     * Value: url to providerListener map
+     */
+    private final ConcurrentHashMap<String, ConcurrentHashMap<Url, ProviderListener>> serviceListeners = new ConcurrentHashMap<>();
+    /**
+     * Key: protocol plus path
+     * Value: url to commandListener map
+     */
+    private final ConcurrentHashMap<String, ConcurrentHashMap<Url, CommandListener>>  commandListeners = new ConcurrentHashMap<>();
 
-    private AbstractConsulClient   client;
-    private ConsulHeartbeatManager heartbeatManager;
-    private int                    lookupInterval;
-
-    // service local cache. key: group, value: <service interface name, url list>
-    private ConcurrentHashMap<String, ConcurrentHashMap<String, List<Url>>> serviceCache = new ConcurrentHashMap<>();
-    // command local cache. key: group, value: command content
-    private ConcurrentHashMap<String, String>                               commandCache = new ConcurrentHashMap<>();
-
-    // record lookup service thread, insure each group start only one thread, <group, lastConsulIndexId>
-    private ConcurrentHashMap<String, Long>   lookupGroupServices = new ConcurrentHashMap<>();
-    // record lookup command thread, <group, command>
-    // TODO: 2016/6/17 change value to consul index
-    private ConcurrentHashMap<String, String> lookupGroupCommands = new ConcurrentHashMap<>();
-
-    // TODO: 2016/6/17 clientUrl support multiple listener
-    // record subscribers service callback listeners, listener was called when corresponding service changes
-    private ConcurrentHashMap<String, ConcurrentHashMap<Url, ProviderListener>> serviceListeners = new ConcurrentHashMap<>();
-    // record subscribers command callback listeners, listener was called when corresponding command changes
-    private ConcurrentHashMap<String, ConcurrentHashMap<Url, CommandListener>>  commandListeners = new ConcurrentHashMap<>();
-    private ThreadPoolExecutor                                                  notifyExecutor;
-
-    public ConsulRegistry(Url url, AbstractConsulClient client) {
-        super(url);
-        this.client = client;
-
-        heartbeatManager = new ConsulHeartbeatManager(client);
-        heartbeatManager.start();
-//        lookupInterval = super.registryUrl.getIntOption(URLParamType.registrySessionTimeout.getName(), DEFAULT_LOOKUP_INTERVAL);
-        lookupInterval = DEFAULT_LOOKUP_INTERVAL;
-
-        ArrayBlockingQueue<Runnable> workQueue = new ArrayBlockingQueue<Runnable>(20000);
-        notifyExecutor = new ThreadPoolExecutor(10, 30, 30 * 1000, TimeUnit.MILLISECONDS, workQueue);
+    public ConsulRegistry(Url registryUrl, LuixConsulClient consulClient) {
+        super(registryUrl);
+        this.consulClient = consulClient;
+        consulStatusUpdater = new ConsulStatusUpdater(consulClient);
+        consulStatusUpdater.start();
+        discoverInterval = this.registryUrl.getIntOption(DISCOVERY_INTERVAL, DISCOVERY_INTERVAL_VAL_DEFAULT);
+        notificationThreadPool = createNotificationThreadPool();
         ShutdownHook.add(this);
-        log.info("ConsulRegistry init finish.");
+        log.info("Initialized consul registry");
+    }
+
+    private ThreadPoolExecutor createNotificationThreadPool() {
+        return new ThreadPoolExecutor(10, 30, 30 * 1_000,
+                TimeUnit.MILLISECONDS, createWorkQueue(), new ThreadPoolExecutor.AbortPolicy());
+    }
+
+    private BlockingQueue<Runnable> createWorkQueue() {
+        return new ArrayBlockingQueue<>(20_000);
     }
 
     @Override
-    protected void doRegister(Url url) {
-        ConsulService service = ConsulUtils.buildService(url);
-        client.registerService(service);
-        heartbeatManager.addHeartbeatServcieId(service.getId());
+    protected void doRegister(Url providerUrl) {
+        ConsulService service = ConsulService.byProviderUrl(providerUrl);
+        consulClient.registerService(service);
     }
 
     @Override
-    protected void doUnregister(Url url) {
-        ConsulService service = ConsulUtils.buildService(url);
-        client.unregisterService(service.getId());
-        heartbeatManager.removeHeartbeatServiceId(service.getId());
+    protected void doDeregister(Url providerUrl) {
+        ConsulService service = ConsulService.byProviderUrl(providerUrl);
+        consulClient.deregisterService(service.getInstanceId());
     }
 
     @Override
-    protected void doActivate(Url url) {
-        if (url == null) {
-            heartbeatManager.setHeartbeatOpen(true);
+    protected void doActivate(Url providerUrl) {
+        if (providerUrl == null) {
+            // Activate all service instances
+            consulStatusUpdater.updateStatus(true);
         } else {
-            throw new UnsupportedOperationException("Command consul registry not support available by urls yet");
+            // Activate specified service instance
+            consulStatusUpdater.activate(ConsulUtils.buildServiceInstanceId(providerUrl));
         }
     }
 
     @Override
-    protected void doDeactivate(Url url) {
-        if (url == null) {
-            heartbeatManager.setHeartbeatOpen(false);
+    protected void doDeactivate(Url providerUrl) {
+        if (providerUrl == null) {
+            // Deactivate all service instances
+            consulStatusUpdater.updateStatus(false);
         } else {
-            throw new UnsupportedOperationException("Command consul registry not support unavailable by urls yet");
+            // Deactivate specified service instance
+            consulStatusUpdater.deactivate(ConsulUtils.buildServiceInstanceId(providerUrl));
         }
     }
 
     @Override
     protected List<Url> discoverActiveProviders(Url consumerUrl) {
-        String service = ConsulUtils.getUrlClusterInfo(consumerUrl);
-        String group = consumerUrl.getForm();
-        List<Url> serviceUrls = new ArrayList<Url>();
-        ConcurrentHashMap<String, List<Url>> serviceMap = serviceCache.get(group);
-        if (serviceMap == null) {
-            synchronized (group.intern()) {
-                serviceMap = serviceCache.get(group);
-                if (serviceMap == null) {
-                    ConcurrentHashMap<String, List<Url>> groupUrls = lookupServiceUpdate(group);
-                    updateServiceCache(group, groupUrls, false);
-                    serviceMap = serviceCache.get(group);
+        String protocolPlusPath = ConsulUtils.getProtocolPlusPath(consumerUrl);
+        String form = consumerUrl.getForm();
+        List<Url> providerUrls = new ArrayList<>();
+        ConcurrentHashMap<String, List<Url>> protocolPlusPath2Urls = urlCache.get(form);
+        if (protocolPlusPath2Urls == null) {
+            synchronized (form.intern()) {
+                protocolPlusPath2Urls = urlCache.get(form);
+                if (protocolPlusPath2Urls == null) {
+                    ConcurrentHashMap<String, List<Url>> protocolPlusPath2UrlsMap = doDiscoverActiveProviders(form);
+                    updateProviderUrlsCache(form, protocolPlusPath2UrlsMap, false);
+                    protocolPlusPath2Urls = urlCache.get(form);
                 }
             }
         }
-        if (serviceMap != null) {
-            serviceUrls = serviceMap.get(service);
+        if (protocolPlusPath2Urls != null) {
+            providerUrls = protocolPlusPath2Urls.get(protocolPlusPath);
         }
-        return serviceUrls;
+        return providerUrls;
     }
 
-    private ConcurrentHashMap<String, List<Url>> lookupServiceUpdate(String group) {
-        ConcurrentHashMap<String, List<Url>> groupUrls = new ConcurrentHashMap<String, List<Url>>();
-        Long lastConsulIndexId = lookupGroupServices.get(group) == null ? 0L : lookupGroupServices.get(group);
-        ConsulResponse<List<ConsulService>> response = lookupConsulService(group, lastConsulIndexId);
+    private ConcurrentHashMap<String, List<Url>> doDiscoverActiveProviders(String form) {
+        ConcurrentHashMap<String, List<Url>> protocolPlusPath2Urls = new ConcurrentHashMap<>();
+        Long lastConsulIndexId = form2ConsulIndex.get(form) == null ? 0L : form2ConsulIndex.get(form);
+        Response<List<ConsulService>> response = consulClient
+                .queryActiveServiceInstances(ConsulUtils.buildProviderServiceName(form), lastConsulIndexId);
         if (response != null) {
-            List<ConsulService> services = response.getValue();
-            if (services != null && !services.isEmpty()
-                    && response.getConsulIndex() > lastConsulIndexId) {
-                for (ConsulService service : services) {
+            List<ConsulService> activeServiceInstances = response.getValue();
+            if (CollectionUtils.isNotEmpty(activeServiceInstances) && response.getConsulIndex() > lastConsulIndexId) {
+                for (ConsulService activeServiceInstance : activeServiceInstances) {
                     try {
-                        Url url = ConsulUtils.buildUrl(service);
-                        String cluster = ConsulUtils.getUrlClusterInfo(url);
-                        List<Url> urlList = groupUrls.get(cluster);
-                        if (urlList == null) {
-                            urlList = new ArrayList<Url>();
-                            groupUrls.put(cluster, urlList);
-                        }
-                        urlList.add(url);
+                        Url url = ConsulUtils.buildUrl(activeServiceInstance);
+                        String protocolPlusPath = ConsulUtils.getProtocolPlusPath(url);
+                        List<Url> urls = protocolPlusPath2Urls.computeIfAbsent(protocolPlusPath, k -> new ArrayList<>());
+                        urls.add(url);
                     } catch (Exception e) {
-                        log.error("convert consul service to url fail! service:" + service, e);
+                        log.error("Failed to build url from consul service instance: " + activeServiceInstance, e);
                     }
                 }
-                lookupGroupServices.put(group, response.getConsulIndex());
-                return groupUrls;
+                form2ConsulIndex.put(form, response.getConsulIndex());
+                return protocolPlusPath2Urls;
             } else {
-                log.info(group + " no need update, lastIndex:" + lastConsulIndexId);
+                log.info("No active service found for form: [{}]", form);
             }
         }
-        return groupUrls;
+        return protocolPlusPath2Urls;
     }
 
-    /**
-     * update service cache of the group.
-     * update local cache when service list changed,
-     * if need notify, notify service
-     *
-     * @param group
-     * @param groupUrls
-     * @param needNotify
-     */
-    private void updateServiceCache(String group, ConcurrentHashMap<String, List<Url>> groupUrls, boolean needNotify) {
-        if (groupUrls != null && !groupUrls.isEmpty()) {
-            ConcurrentHashMap<String, List<Url>> groupMap = serviceCache.get(group);
-            if (groupMap == null) {
-                serviceCache.put(group, groupUrls);
-            }
-            for (Map.Entry<String, List<Url>> entry : groupUrls.entrySet()) {
+    private void updateProviderUrlsCache(String form, ConcurrentHashMap<String, List<Url>> protocolPlusPath2Urls, boolean needNotify) {
+        if (MapUtils.isNotEmpty(protocolPlusPath2Urls)) {
+            ConcurrentHashMap<String, List<Url>> protocolPlusPath2UrlsCopy = urlCache.putIfAbsent(form, protocolPlusPath2Urls);
+            for (Map.Entry<String, List<Url>> entry : protocolPlusPath2Urls.entrySet()) {
                 boolean change = true;
-                if (groupMap != null) {
-                    List<Url> oldUrls = groupMap.get(entry.getKey());
+                if (protocolPlusPath2UrlsCopy != null) {
+                    List<Url> oldUrls = protocolPlusPath2UrlsCopy.get(entry.getKey());
                     List<Url> newUrls = entry.getValue();
-                    if (newUrls == null || newUrls.isEmpty() || ConsulUtils.isSame(entry.getValue(), oldUrls)) {
+                    if (CollectionUtils.isEmpty(newUrls) || ConsulUtils.isSame(entry.getValue(), oldUrls)) {
                         change = false;
                     } else {
-                        groupMap.put(entry.getKey(), newUrls);
+                        protocolPlusPath2UrlsCopy.put(entry.getKey(), newUrls);
                     }
                 }
                 if (change && needNotify) {
-                    notifyExecutor.execute(new NotifyService(entry.getKey(), entry.getValue()));
-                    log.info("motan service notify-service: " + entry.getKey());
+                    notificationThreadPool.execute(new NotifyService(entry.getKey(), entry.getValue()));
+                    log.info("service notify-service: " + entry.getKey());
                     StringBuilder sb = new StringBuilder();
                     for (Url url : entry.getValue()) {
                         sb.append(url.getUri()).append(";");
                     }
-                    log.info("consul notify urls:" + sb.toString());
+                    log.info("consul notify urls:" + sb);
                 }
             }
         }
-    }
-
-    /**
-     * directly fetch consul service data.
-     *
-     * @param serviceName
-     * @return ConsulResponse or null
-     */
-    private ConsulResponse<List<ConsulService>> lookupConsulService(String serviceName, Long lastConsulIndexId) {
-        ConsulResponse<List<ConsulService>> response = client.lookupHealthService(
-                ConsulUtils.convertGroupToServiceName(serviceName),
-                lastConsulIndexId);
-        return response;
     }
 
     @Override
@@ -209,42 +213,36 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
         startListenerThreadIfNewService(consumerUrl);
     }
 
-    private void addServiceListener(Url url, ProviderListener serviceListener) {
-        String service = ConsulUtils.getUrlClusterInfo(url);
-        ConcurrentHashMap<Url, ProviderListener> map = serviceListeners.get(service);
+    private void addServiceListener(Url url, ProviderListener providerListener) {
+        String protocolPlusPath = ConsulUtils.getProtocolPlusPath(url);
+        ConcurrentHashMap<Url, ProviderListener> map = serviceListeners.get(protocolPlusPath);
         if (map == null) {
-            serviceListeners.putIfAbsent(service, new ConcurrentHashMap<Url, ProviderListener>());
-            map = serviceListeners.get(service);
+            serviceListeners.putIfAbsent(protocolPlusPath, new ConcurrentHashMap<>());
+            map = serviceListeners.get(protocolPlusPath);
         }
         synchronized (map) {
-            map.put(url, serviceListener);
+            map.put(url, providerListener);
         }
     }
 
-    /**
-     * if new group registed, start a new lookup thread
-     * each group start a lookup thread to discover service
-     *
-     * @param url
-     */
     private void startListenerThreadIfNewService(Url url) {
-        String group = url.getForm();
-        if (!lookupGroupServices.containsKey(group)) {
-            Long value = lookupGroupServices.putIfAbsent(group, 0L);
-            if (value == null) {
-                ServiceLookupThread lookupThread = new ServiceLookupThread(group);
-                lookupThread.setDaemon(true);
-                lookupThread.start();
+        if (!form2ConsulIndex.containsKey(url.getForm())) {
+            Long consulIndex = form2ConsulIndex.putIfAbsent(url.getForm(), 0L);
+            if (consulIndex == null) {
+                DiscoverProviderThread discoverProviderThread = new DiscoverProviderThread(url.getForm());
+                discoverProviderThread.setDaemon(true);
+                discoverProviderThread.start();
             }
         }
     }
 
     @Override
     protected void unsubscribeProviderListener(Url consumerUrl, ProviderListener listener) {
-        ConcurrentHashMap<Url, ProviderListener> listeners = serviceListeners.get(ConsulUtils.getUrlClusterInfo(consumerUrl));
-        if (listeners != null) {
-            synchronized (listeners) {
-                listeners.remove(consumerUrl);
+        ConcurrentHashMap<Url, ProviderListener> url2ProviderListeners =
+                serviceListeners.get(ConsulUtils.getProtocolPlusPath(consumerUrl));
+        if (url2ProviderListeners != null) {
+            synchronized (url2ProviderListeners) {
+                url2ProviderListeners.remove(consumerUrl);
             }
         }
     }
@@ -269,8 +267,8 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
 
     private void startListenerThreadIfNewCommand(Url url) {
         String group = url.getForm();
-        if (!lookupGroupCommands.containsKey(group)) {
-            String command = lookupGroupCommands.putIfAbsent(group, "");
+        if (!form2Command.containsKey(group)) {
+            String command = form2Command.putIfAbsent(group, StringUtils.EMPTY);
             if (command == null) {
                 CommandLookupThread lookupThread = new CommandLookupThread(group);
                 lookupThread.setDaemon(true);
@@ -290,6 +288,31 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
     }
 
     @Override
+    public void subscribe(Url consumerUrl) {
+        ConsulService service = ConsulService.byConsumerUrl(consumerUrl);
+        consulClient.registerService(service);
+        // Activate specified service instance
+        consulStatusUpdater.activate(service.getInstanceId());
+    }
+
+    @Override
+    public void unsubscribe(Url consumerUrl) {
+        ConsulService service = ConsulService.byConsumerUrl(consumerUrl);
+        consulClient.deregisterService(service.getInstanceId());
+    }
+
+    @Override
+    public List<Url> getAllProviderUrls() {
+        return consulClient.getAllProviderUrls();
+    }
+
+    @Override
+    public void subscribeConsumerListener(String interfaceName, ConsumerProcessable consumerProcessor) {
+        List<Url> consumerUrls = consulClient.getConsumerUrls(interfaceName);
+        consumerProcessor.process(getRegistryUrl(), interfaceName, consumerUrls);
+    }
+
+    @Override
     protected String readCommand(Url consumerUrl) {
         String group = consumerUrl.getForm();
         String command = lookupCommandUpdate(group);
@@ -297,29 +320,9 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
         return command;
     }
 
-    @Override
-    public List<String> getAllProviderPaths() {
-        return null;
-    }
-
-    @Override
-    public List<String> discoverActiveProviderAddress(String providerPath) {
-        return null;
-    }
-
-    @Override
-    public void subscribeConsumerListener(String interfaceName, ConsumerProcessable consumerProcessor) {
-
-    }
-
-    @Override
-    public void cleanup() {
-        heartbeatManager.close();
-    }
-
     private String lookupCommandUpdate(String group) {
-        String command = client.lookupCommand(group);
-        lookupGroupCommands.put(group, command);
+        String command = consulClient.queryCommand(group);
+        form2Command.put(group, command);
         return command;
     }
 
@@ -327,17 +330,13 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
      * update command cache of the group.
      * update local cache when command changed,
      * if need notify, notify command
-     *
-     * @param group
-     * @param command
-     * @param needNotify
      */
     private void updateCommandCache(String group, String command, boolean needNotify) {
         String oldCommand = commandCache.get(group);
         if (!command.equals(oldCommand)) {
             commandCache.put(group, command);
             if (needNotify) {
-                notifyExecutor.execute(new NotifyCommand(group, command));
+                notificationThreadPool.execute(new NotifyCommand(group, command));
                 log.info(String.format("command data change: group=%s, command=%s: ", group, command));
             }
         } else {
@@ -350,8 +349,8 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
     }
 
     private class NotifyService implements Runnable {
-        private String    service;
-        private List<Url> urls;
+        private final String    service;
+        private final List<Url> urls;
 
         public NotifyService(String service, List<Url> urls) {
             this.service = service;
@@ -375,17 +374,17 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
     }
 
     private class NotifyCommand implements Runnable {
-        private String group;
-        private String command;
+        private final String form;
+        private final String command;
 
-        public NotifyCommand(String group, String command) {
-            this.group = group;
+        public NotifyCommand(String form, String command) {
+            this.form = form;
             this.command = command;
         }
 
         @Override
         public void run() {
-            ConcurrentHashMap<Url, CommandListener> listeners = commandListeners.get(group);
+            ConcurrentHashMap<Url, CommandListener> listeners = commandListeners.get(form);
             synchronized (listeners) {
                 for (Map.Entry<Url, CommandListener> entry : listeners.entrySet()) {
                     CommandListener commandListener = entry.getValue();
@@ -395,25 +394,25 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
         }
     }
 
-    private class ServiceLookupThread extends Thread {
-        private String group;
+    private class DiscoverProviderThread extends Thread {
+        private final String form;
 
-        public ServiceLookupThread(String group) {
-            this.group = group;
+        public DiscoverProviderThread(String form) {
+            this.form = form;
         }
 
         @Override
         public void run() {
-            log.info("start group lookup thread. lookup interval: " + lookupInterval + "ms, group: " + group);
+            log.info("Start discover providers thread with interval: " + discoverInterval + "ms and form: " + form);
             while (true) {
                 try {
-                    sleep(lookupInterval);
-                    ConcurrentHashMap<String, List<Url>> groupUrls = lookupServiceUpdate(group);
-                    updateServiceCache(group, groupUrls, true);
+                    sleep(discoverInterval);
+                    ConcurrentHashMap<String, List<Url>> protocolPlusPath2Urls = doDiscoverActiveProviders(form);
+                    updateProviderUrlsCache(form, protocolPlusPath2Urls, true);
                 } catch (Throwable e) {
-                    log.error("group lookup thread fail!", e);
+                    log.error("Failed to discover providers!", e);
                     try {
-                        Thread.sleep(2000);
+                        Thread.sleep(2_000);
                     } catch (InterruptedException ignored) {
                     }
                 }
@@ -422,7 +421,7 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
     }
 
     private class CommandLookupThread extends Thread {
-        private String group;
+        private final String group;
 
         public CommandLookupThread(String group) {
             this.group = group;
@@ -430,10 +429,10 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
 
         @Override
         public void run() {
-            log.info("start command lookup thread. lookup interval: " + lookupInterval + "ms, group: " + group);
+            log.info("start command lookup thread. lookup interval: " + discoverInterval + "ms, group: " + group);
             while (true) {
                 try {
-                    sleep(lookupInterval);
+                    sleep(discoverInterval);
                     String command = lookupCommandUpdate(group);
                     updateCommandCache(group, command, true);
                 } catch (Throwable e) {
@@ -445,5 +444,12 @@ public class ConsulRegistry extends CommandFailbackAbstractRegistry implements C
                 }
             }
         }
+    }
+
+    @Override
+    public void destroy() {
+        notificationThreadPool.shutdown();
+        consulStatusUpdater.close();
+        log.info("Destroyed consul registry");
     }
 }
