@@ -10,7 +10,6 @@ import java.util.List;
 
 /**
  * A utility used to clean up the resources
- *
  */
 @Slf4j
 @ThreadSafe
@@ -34,18 +33,18 @@ public class ShutdownHook extends Thread {
     private ShutdownHook() {
     }
 
-    public static synchronized void add(Cleanable cleanable, int priority) {
-        RESOURCES.add(new CleanableObject(cleanable, priority));
-        log.info("Added the cleanup method of class [{}] to {}", cleanable.getClass().getSimpleName(), ShutdownHook.class.getSimpleName());
+    public static synchronized void add(Destroyable destroyable, int priority) {
+        RESOURCES.add(new CleanableObject(destroyable, priority));
+        log.info("Added the cleanup method of class [{}] to {}", destroyable.getClass().getSimpleName(), ShutdownHook.class.getSimpleName());
     }
 
     /**
      * Only global resources are allowed to add to it.
      *
-     * @param cleanable cleanable
+     * @param destroyable cleanable
      */
-    public static void add(Cleanable cleanable) {
-        add(cleanable, DEFAULT_PRIORITY);
+    public static void add(Destroyable destroyable) {
+        add(destroyable, DEFAULT_PRIORITY);
     }
 
     /**
@@ -53,6 +52,7 @@ public class ShutdownHook extends Thread {
      */
     public static void register() {
         Runtime.getRuntime().addShutdownHook(INSTANCE);
+        log.info("Registered the {} to system runtime", ShutdownHook.class.getSimpleName());
     }
 
     public static void runNow(boolean sync) {
@@ -83,19 +83,20 @@ public class ShutdownHook extends Thread {
         Collections.sort(RESOURCES);
         for (CleanableObject resource : RESOURCES) {
             try {
-                resource.cleanable.cleanup();
+                resource.destroyable.destroy();
             } catch (Exception e) {
-                log.error("Failed to cleanup " + resource.cleanable.getClass().getSimpleName(), e);
+                System.out.println("Failed to cleaned up the " + resource.destroyable.getClass().getSimpleName()
+                        + " by ShutdownHook with exception: " + e.getMessage());
             }
-            log.info("Cleaned up the {}", resource.cleanable.getClass().getSimpleName());
+            System.out.println("Cleaned up the " + resource.destroyable.getClass().getSimpleName() + " by ShutdownHook");
         }
         RESOURCES.clear();
     }
 
     @AllArgsConstructor
     private static class CleanableObject implements Comparable<CleanableObject> {
-        private final Cleanable cleanable;
-        private final int       priority;
+        private final Destroyable destroyable;
+        private final int         priority;
 
         /**
          * Lower values have higher priority
