@@ -12,5 +12,5 @@ public final class ApplicationConstants {
     public static final String   SPRING_PROFILE_PROD = "prod";
     public static final String[] AVAILABLE_PROFILES  = new String[]{SPRING_PROFILE_DEV, SPRING_PROFILE_DEMO, SPRING_PROFILE_PROD};
     public static final Locale   SYSTEM_LOCALE       = Locale.US;
-    public static final String   DEFAULT_REG         = "zookeeper://localhost:2181/registry";
+    public static final String   DEFAULT_REG         = "consul://localhost:8500/registry";
 }
